refactor(index): move config logic into ConfigBar
